Separate out types.
