+++ README.md
@@ -31,9 +31,12 @@ The API takes inspiration from Apple's [`GestureEvent` specification](https://de
 ```javascript
 const myDiv = document.querySelector('#myDiv');
 const unregister = twoFingers(myDiv, {
-  gestureStart: (g) => console.log(`Gesture start. Gesture: ${g.toString()}`),
-  gestureChange: (g) => console.log(`Gesture change. Gesture: ${g.toString()}`),
-  gestureEnd: (g) => console.log(`Gesture end. Gesture: ${g.toString()}`),
+  onGestureStart: (g) => console.log(`Gesture start. Gesture: ${g.toString()}`),
+  onGestureChange: (g) => console.log(`Gesture change. Gesture: ${g.toString()}`),
+  onGestureEnd: (g) => console.log(`Gesture end. Gesture: ${g.toString()}`),
+}, {
+  scaleSpeedup: 0.25,
+  translationSpeedUp: 1,
 });
 ```
